pyproject.toml
[project]
name = "code-display"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "py-cord>=2.6.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]
